Fix: redis tests
