+++ front/lib/api/redis.ts
@@ -40,7 +40,7 @@ export async function getRedisClient({
       isolationPoolOptions: {
         acquireTimeoutMillis: 10000, // Max time to wait for a connection: 10 seconds.
         min: 1,
-        max: 800, // Maximum number of concurrent connections for streaming.
+        max: 8000, // Maximum number of concurrent connections for streaming.
         evictionRunIntervalMillis: 15000, // Check for idle connections every 15 seconds.
         idleTimeoutMillis: 30000, // Connections idle for more than 30 seconds will be eligible for eviction.
       },
